requirements.txt
telebot
dotenv
transformers
torch
openai-whisper
accelerate 
sentencepiece
